preparing for code-reorg: data boundaries will be calculated by CartesianDiagram, with Bar~/LineDiagram only doing slight adjusting afterwards (in a seperate mthod, receiving raw boundaries amd first/last/max threeD depth)    Note: current code only implemented half, do not try out.
git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/trunk/kdchart@1202 c66179b2-ff0d-0410-8aef-1b1a91e65cfa

+++ src/KDChartAbstractThreeDAttributes_p.h
@@ -57,7 +57,7 @@ public:
 private:
     bool enabled;
     int height;
-    int depth;
+    double depth;
 };
 
 inline AbstractThreeDAttributes::AbstractThreeDAttributes( Private * p ) : _d( p ) { init(); }
